+++ integration-tests/src/test/java/arez/integration/DaggerRepositoryIntegrationTest.java
@@ -3,7 +3,7 @@ package arez.integration;
 import arez.Arez;
 import arez.ArezContext;
 import arez.annotations.ArezComponent;
-import arez.annotations.Injectible;
+import arez.annotations.Feature;
 import arez.annotations.Observable;
 import arez.annotations.Repository;
 import dagger.Component;
@@ -14,7 +14,7 @@ import static org.testng.Assert.*;
 public class DaggerRepositoryIntegrationTest
 {
   @ArezComponent
-  @Repository( dagger = Injectible.ENABLE )
+  @Repository( dagger = Feature.ENABLE )
   static class TestComponent
   {
     private String _value;