+++ processor/src/main/java/arez/processor/GeneratorUtil.java
@@ -21,7 +21,7 @@ final class GeneratorUtil
   static final ClassName COMPUTED_VALUE_CLASSNAME = ClassName.get( "arez", "ComputedValue" );
   static final ClassName DISPOSABLE_CLASSNAME = ClassName.get( "arez", "Disposable" );
   static final ClassName COMPONENT_CLASSNAME = ClassName.get( "arez", "Component" );
-  static final ClassName INJECTIBLE_CLASSNAME = ClassName.get( "arez.annotations", "Injectible" );
+  static final ClassName INJECTIBLE_CLASSNAME = ClassName.get( "arez.annotations", "Feature" );
   static final ClassName NO_RESULT_EXCEPTION_CLASSNAME =
     ClassName.get( "arez.component", "NoResultException" );
   static final ClassName NO_SUCH_ENTITY_EXCEPTION_CLASSNAME =