+++ processor/src/test/resources/input/com/example/repository/DaggerEnabledRepository.java
@@ -1,12 +1,12 @@
 package com.example.repository;
 
 import arez.annotations.ArezComponent;
-import arez.annotations.Injectible;
+import arez.annotations.Feature;
 import arez.annotations.Observable;
 import arez.annotations.Repository;
 import javax.annotation.Nonnull;
 
-@Repository( dagger = Injectible.ENABLE )
+@Repository( dagger = Feature.ENABLE )
 @ArezComponent
 public class DaggerEnabledRepository
 {